+++ frontend/src/pages/protocol-token/index.vue
@@ -6,59 +6,87 @@
 			<div
 				class="flex flex-row justify-center lg:place-content-end mb-4 lg:mb-0"
 			></div>
-			<!-- <header class="flex justify-between items-center mb-4">
+			<header class="flex justify-between items-center mb-4">
 				<h1 class="text-xl">Overview</h1>
 			</header>
 			<FtbCarousel non-carousel-classes="grid-cols-4">
 				<CarouselSlide class="w-full lg:h-full">
-					<KeyValueChartCard class="w-96 lg:w-full" :y-values="[[null]]" :is-loading="false">
-						<template #title>Total Market Cap</template>
+					<KeyValueChartCard
+						class="w-96 lg:w-full"
+						:y-values="[[null]]"
+						:is-loading="false"
+					>
+						<template #title>Total Token Supply</template>
+						<template #value>
+							<p class="font-bold text-2xl mt-2">
+								<PLtAmount
+									:value="
+										pltTokenDataRef
+											.reduce(
+												(acc, coin) =>
+													acc +
+													calculateActualValue(
+														String(coin?.totalSupply),
+														Number(coin?.decimal)
+													),
+												0n
+											)
+											.toString()
+									"
+									:decimals="0"
+									:format-number="true"
+								/>
+							</p>
+						</template>
+					</KeyValueChartCard>
+				</CarouselSlide>
+				<CarouselSlide class="w-full lg:h-full">
+					<KeyValueChartCard
+						class="w-96 lg:w-full"
+						:y-values="[[null]]"
+						:is-loading="pltEventMetricsLoading"
+					>
+						<template #title>Unique Holders</template>
+						<template #value>
+							<p class="font-bold text-2xl mt-2">
+								{{ pltUniqueAccountsDataRef?.pltUniqueAccounts }}
+							</p>
+						</template>
+					</KeyValueChartCard>
+				</CarouselSlide>
+				<CarouselSlide class="w-full lg:h-full">
+					<KeyValueChartCard
+						class="w-96 lg:w-full"
+						:y-values="[[null]]"
+						:is-loading="pltEventMetricsLoading"
+					>
+						<template #title># of Txs (24h)</template>
+						<template #value>
+							<p class="font-bold text-2xl mt-2">
+								{{ pltEventMetricsDataRef?.pltEventMetrics.totalEventCount }}
+							</p>
+						</template>
+					</KeyValueChartCard>
+				</CarouselSlide>
+				<CarouselSlide class="w-full lg:h-full">
+					<KeyValueChartCard
+						class="w-96 lg:w-full"
+						:y-values="[[null]]"
+						:is-loading="pltEventMetricsLoading"
+					>
+						<template #title>Total Transfer Volume (24h)</template>
 						<template #value>
 							<p class="font-bold text-2xl mt-2">
 								{{
-									
 									numberFormatter(
-										pltEventMetricsDataRef?.pltEventMetrics.lastCumulativeTotalSupply
+										pltEventMetricsDataRef?.pltEventMetrics.transferVolume
 									)
 								}}
 							</p>
 						</template>
 					</KeyValueChartCard>
 				</CarouselSlide>
-				<CarouselSlide class="w-full lg:h-full">
-					<KeyValueChartCard class="w-96 lg:w-full" :y-values="[[null]]" :is-loading="false">
-						<template #title>Unique Holders</template>
-						<template #value>
-							<p class="font-bold text-2xl mt-2">
-								{{ pltEventMetricsDataRef?.pltEventMetrics.lastCumulativeUniqueHolders}}
-							</p>
-						</template>
-					</KeyValueChartCard>
-				</CarouselSlide>
-				<CarouselSlide class="w-full lg:h-full">
-					<KeyValueChartCard class="w-96 lg:w-full" :y-values="[[null]]" :is-loading="false">
-						<template #title># of Txs (24h)</template>
-						<template #value>
-							<p class="font-bold text-2xl mt-2">
-								{{ pltEventMetricsDataRef?.pltEventMetrics.eventCount }}
-							</p>
-						</template>
-					</KeyValueChartCard>
-				</CarouselSlide>
-				<CarouselSlide class="w-full lg:h-full">
-					<KeyValueChartCard class="w-96 lg:w-full" :y-values="[[null]]" :is-loading="false">
-						<template #title>Total Values Transfer (24h)</template>
-						<template #value>
-							<p class="font-bold text-2xl mt-2">
-								{{
-									'$' +
-									0.0
-								}}
-							</p>
-						</template>
-					</KeyValueChartCard>
-				</CarouselSlide>
-			</FtbCarousel> -->
+			</FtbCarousel>
 			<header class="flex justify-between items-center mb-4">
 				<h1 class="text-xl">Supply & Holders</h1>
 			</header>
@@ -119,13 +147,15 @@
 						</TableTd>
 
 						<TableTd>
-							<!-- <a :href="`/protocol-token/${event.tokenId}`" target="_blank"
-								class="font-normal text-md text-theme-interactive flex flex-row items-center">
-								<img :src="" class="rounded-full w-6 h-6 mr-2"
-									alt="Token Icon" loading="lazy" decoding="async" />
-							
-							</a> -->
-							{{ event.tokenName }}
+							<a
+								:href="`/protocol-token/${event.tokenId}`"
+								target="_blank"
+								class="font-normal text-md text-theme-interactive flex flex-row items-center"
+							>
+								{{ event.tokenName }}
+								<!-- <img :src="event.tokenIconUrl" class="rounded-full w-6 h-6 mr-2"
+									alt="Token Icon" loading="lazy" decoding="async" /> -->
+							</a>
 						</TableTd>
 						<TableTd>
 							<AccountLink
@@ -179,19 +209,24 @@
 	</div>
 </template>
 <script lang="ts" setup>
-import { ref } from 'vue'
+import { ref, watch } from 'vue'
 
 import FtbCarousel from '~/components/molecules/FtbCarousel.vue'
 import CarouselSlide from '~/components/molecules/CarouselSlide.vue'
 import StableCoinDistributionChart from '~/components/molecules/ChartCards/StableCoinDistributionChart.vue'
 import StableCoinSupplyBarChart from '~/components/molecules/ChartCards/StableCoinSupplyBarChart.vue'
-import HolderByStableCoin from '~/components/molecules/ChartCards/HolderByStableCoin.vue'
-import { usePltTokenQuery } from '~/queries/usePltTokenQuery'
+import {
+	usePltTokenQuery,
+	usePltUniqueAccountsQuery,
+} from '~/queries/usePltTokenQuery'
 import { usePagedData } from '~/composables/usePagedData'
 import { usePltEventsQuery } from '~/queries/usePltEventsQuery'
 import type { Pltevent } from '~/types/generated'
 import { useDateNow } from '~/composables/useDateNow'
-
+import HolderByStableCoin from '~/components/molecules/ChartCards/HolderByStableCoin.vue'
+import KeyValueChartCard from '~/components/molecules/KeyValueChartCard.vue'
+import { usePltEventsMetricsQuery } from '~/queries/usePltEventsMetricsQuery'
+import { MetricsPeriod } from '~/types/generated'
 definePageMeta({
 	middleware: 'plt-features-guard',
 })
@@ -207,6 +242,7 @@ const { data: pltTokenData, loading: pltTokenLoading } = usePltTokenQuery()
 const { NOW } = useDateNow()
 
 const pltTokenDataRef = ref(pltTokenData)
+const selectedMetricsPeriod = ref(MetricsPeriod.Last24Hours)
 
 watch(
 	pltTokenData,
@@ -228,6 +264,28 @@ watch(
 	() => pltEventsDataRef.value,
 	value => {
 		addPagedData(value?.pltEvents.nodes || [], value?.pltEvents.pageInfo)
+	},
+	{ immediate: true, deep: true }
+)
+
+const { data: pltEventMetricsData, loading: pltEventMetricsLoading } =
+	usePltEventsMetricsQuery(selectedMetricsPeriod)
+const pltEventMetricsDataRef = ref(pltEventMetricsData)
+
+watch(
+	pltEventMetricsData,
+	newData => {
+		pltEventMetricsDataRef.value = newData
+	},
+	{ immediate: true, deep: true }
+)
+
+const { data: pltUniqueAccountsData } = usePltUniqueAccountsQuery()
+const pltUniqueAccountsDataRef = ref(pltUniqueAccountsData)
+watch(
+	pltUniqueAccountsDataRef,
+	newData => {
+		pltUniqueAccountsDataRef.value = newData
 	},
 	{ immediate: true, deep: true }
 )